+++ packages/amazon-bedrock/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/amazon-bedrock
 
+## 3.0.94
+
+### Patch Changes
+
+- 4b40033: feat(provider/anthropic): add support for Opus 4.7 and relevant API enhancements
+- Updated dependencies [4b40033]
+  - @ai-sdk/anthropic@2.0.75
+
 ## 3.0.93
 
 ### Patch Changes