+++ packages/gateway/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @ai-sdk/gateway
 
+## 2.0.81
+
+### Patch Changes
+
+- dc91bcc: Backport: chore(provider/gateway): update gateway model settings files
+
+## 2.0.80
+
+### Patch Changes
+
+- 4b40033: feat(provider/anthropic): add support for Opus 4.7 and relevant API enhancements
+
+## 2.0.79
+
+### Patch Changes
+
+- d4fbea5: feat (provider/gateway): add sort options
+
+## 2.0.78
+
+### Patch Changes
+
+- 140640d: feat (provider/gateway): make model list resilient to unknown model types
+
+## 2.0.77
+
+### Patch Changes
+
+- 6c0cde5: Backport: chore(provider/gateway): update gateway model settings files
+
+## 2.0.76
+
+### Patch Changes
+
+- a5317d0: Backport: chore(provider/gateway): update gateway model settings files
+
+## 2.0.75
+
+### Patch Changes
+
+- f03fec2: Backport: chore(provider/gateway): update gateway model settings files
+
 ## 2.0.74
 
 ### Patch Changes