+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @ai-sdk/google-vertex
 
+## 3.0.130
+
+### Patch Changes
+
+- 4b40033: feat(provider/anthropic): add support for Opus 4.7 and relevant API enhancements
+- Updated dependencies [4b40033]
+  - @ai-sdk/anthropic@2.0.75
+
+## 3.0.129
+
+### Patch Changes
+
+- bfaafad: chore(provider/google): update available models
+- Updated dependencies [bfaafad]
+  - @ai-sdk/google@2.0.69
+
+## 3.0.128
+
+### Patch Changes
+
+- Updated dependencies [4c10694]
+  - @ai-sdk/google@2.0.68
+
 ## 3.0.127
 
 ### Patch Changes