+++ packages/react/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @ai-sdk/react
 
+## 2.0.179
+
+### Patch Changes
+
+- ai@5.0.177
+
+## 2.0.178
+
+### Patch Changes
+
+- ai@5.0.176
+
+## 2.0.177
+
+### Patch Changes
+
+- ai@5.0.175
+
+## 2.0.176
+
+### Patch Changes
+
+- ai@5.0.174
+
+## 2.0.175
+
+### Patch Changes
+
+- ai@5.0.173
+
+## 2.0.174
+
+### Patch Changes
+
+- ai@5.0.172
+
+## 2.0.173
+
+### Patch Changes
+
+- ai@5.0.171
+
 ## 2.0.172
 
 ### Patch Changes