+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,48 @@
 
 ### Patch Changes
 
+- ai@5.0.177
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.176
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.175
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.174
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.173
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.172
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.171
+
+## 0.0.1
+
+### Patch Changes
+
 - ai@5.0.170
 
 ## 0.0.1